% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeguard.R
\name{diff_n}
\alias{diff_n}
\title{Create matrix of differences}
\usage{
diff_n(x, n = 0:1, ...)
}
\arguments{
\item{x}{A vector of values to be lagged.}

\item{n}{An integer vector of at least two values giving the differences to be created. For two values \code{n[1]} and \code{n[2]}, the function creates (\code{x} at time \code{n[2]}) - (\code{x} at time \code{n[1]}).}

\item{...}{Additional arguments, currently not implemented.}
}
\value{
A matrix with the differenced values of \code{x}.
}
\description{
Creates a matrix of differenced values of a variable.
}
\examples{
diff_n(1:10, 0:3)

}
