% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeguard.R
\name{lag_n}
\alias{lag_n}
\title{Create matrix of lags}
\usage{
lag_n(x, n = 1, rep0 = FALSE, ...)
}
\arguments{
\item{x}{A vector of values to be lagged.}

\item{n}{An integer vector giving the lags to be created.  Including 0 will inculde the time t value of \code{x}.}

\item{rep0}{If \code{TRUE}, replaces \code{NA} values with 0.  Default is \code{FALSE}.}

\item{...}{Additional arguments, currently not implemented.}
}
\value{
A matrix with the lagged values of \code{x}.  The first column is the time t value of \code{x} (if \code{0} is included in \code{n}), and subsequent columns are the lagged values.
}
\description{
Creates a matrix of lagged values of a variable.
}
\examples{
lag_n(1:10, 0:3)

}
