% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeguard.R
\name{lrm_ci}
\alias{lrm_ci}
\title{Confidence Intervals for Long-run Multipliers}
\usage{
lrm_ci(
  obj,
  y_coefs,
  x_coefs,
  qb_dist = c("t", "normal"),
  level = 0.95,
  t_min = 2,
  ...
)
}
\arguments{
\item{obj}{A linear regression model in ADL form.}

\item{y_coefs}{A vector of index values for the lag-y coefficients from the model.}

\item{x_coefs}{A list of vectors of index values for the level and lag-x coefficients from the model. Each
element of the list is assumed to be a different variable whose LRM is to be calculated.  If the list is
named, the output uses those names.}

\item{qb_dist}{The distribution for the quasi-Bayesian simulation either \code{"normal"} or \code{"t"} (default is \code{"t"}).}

\item{level}{The confidence level for the intervals.  Default is \code{0.95}.}

\item{t_min}{Minimum adjusted degrees of freedom in the t-distribution simulation.  Default is 2.}

\item{...}{Additional arguments, currently not implemented.}
}
\value{
A data frame with the following variables:
\itemize{
\item \code{vbl} - The name of the variable for which the long-run multiplier is calculated
\item \code{est} - The estimate of the long-run multiplier
\item \code{delta_se} - The approximate standard error of the estimate calculated by the delta method
\item \code{lwr_h}, \code{upr_h} - The confidence bounds calculated from the Hinkley distribution.
\item \code{lwr_qbn}, \code{upr_qbn} - The confidence bounds calculated from a quasi-Bayesian simulation using a multivariate normal distribution.
\item \code{lwr_qbt}, \code{upr_qbt} - The confidence bounds calculated from a quasi-Bayesian simulation using a multivariate t distribution.
\item \code{lwr_t}, \code{upr_t} - The confidence bounds calculated from the t-distribution using the delta method SE
\item \code{lwr_f1}, \code{upr_f1}, \code{lwr_f2}, \code{upr_f2} - The confidence bounds calculated from Fieller's theorem.  If the confidence set is unbounded, both lower and both upper bounds are \code{NA}.  If the confidence interval is bounded or half-line, the \code{lwr_f2} = \code{upr_f2} = \code{NA}.
}
}
\description{
Calculates confidence intervals for long-run multipliers from a time-series analysis.
}
\details{
The function assumes that the object is the autoregressive distributed lag (ADL)
representation of the regression model.  The confidence intervals are made from both the
t-distribution (using the delta method for standard errors) and with the Hinkley distribution.
Assuming \code{b} is the vector of model coefficients, the LRM is calculated with
\verb{sum(b[x_coefs[[i]])/(1-sum(b[y_coefs]))} for each of the \code{i} elements of \code{x_coefs}.
}
\examples{
x <- y <- c(0, rep(NA, 99))
for(t in 2:100){
  x[t] <- .5*x[(t-1)] + rnorm(1,0,1)
}

for(t in 2:100){
y[t] <- .7*y[(t-1)] + .5*x[t] + .25*x[(t-1)] + rnorm(1,0,1)
}

d <- data.frame(time=1:100, x=x, y=y)
adl <- lm(y ~ lag_n(y,1) + lag_n(x, c(0,1)), data=d)
lrm_ci(adl, y_coefs = 2, x_coefs = list(x=3:4))
}
