% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeguard.R
\name{phinkley}
\alias{phinkley}
\title{CDF of Hinkley Distribution}
\usage{
phinkley(r, mu_a, mu_b, sigma_a, sigma_b, sigma_ab, ...)
}
\arguments{
\item{r}{A numeric vector of ratios for which to calculate the CDF.}

\item{mu_a}{The mean of the numerator normal distribution.}

\item{mu_b}{The mean of the denominator normal distribution.}

\item{sigma_a}{The standard deviation of the numerator normal distribution.}

\item{sigma_b}{The standard deviation of the denominator normal distribution.}

\item{sigma_ab}{The covariance between the numerator and denominator normal distributions.}

\item{...}{Additional arguments, currently not implemented.}
}
\value{
A numeric vector of the CDF values for the given ratios.
}
\description{
Calculates the CDF of the Hinkley distribution.  This is particularly useful for the ratio of normal variates.
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100, 0, .5)
v <- var(cbind(x,y))
phinkley(mean(x)/mean(y), mean(x), mean(y), v[1,1], v[2,2], v[1,2])
}
