% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeguard.R
\name{qhinkley}
\alias{qhinkley}
\title{Quantile function for the Hinkley distribution}
\usage{
qhinkley(
  p,
  mu_a,
  mu_b,
  sigma_a,
  sigma_b,
  sigma_ab,
  tol = 1e-08,
  maxiter = 60,
  ...
)
}
\arguments{
\item{p}{A numeric vector of quantile values.}

\item{mu_a}{The mean of the numerator normal distribution.}

\item{mu_b}{The mean of the denominator normal distribution.}

\item{sigma_a}{The standard deviation of the numerator normal distribution.}

\item{sigma_b}{The standard deviation of the denominator normal distribution.}

\item{sigma_ab}{The covariance between the numerator and denominator normal distributions.}

\item{tol}{The tolerance for the root-finding algorithm. Default is \code{1e-8}.}

\item{maxiter}{The maximum number of iterations for the root-finding algorithm. Default is \code{60}.}

\item{...}{Additional arguments, currently not implemented.}
}
\value{
A numeric vector of quantiles corresponding to the input probabilities.
}
\description{
Estimates the quantiles of the Hinkley distribution.
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100, 0, .5)
v <- var(cbind(x,y))
qhinkley(.975, 0, mean(y), v[1,1], v[2,2], v[1,2])
}
